requirements.txt
Flask==3.1.2
flask-login==0.6.3
Pillow==11.3.0
werkzeug==3.1.3
gunicorn==22.0.0
flask-wtf==1.2.2
openai
google-genai
bleach
google-genai
sift-stack-py
google-genai